requirements.txt
asn1crypto==0.24.0
certifi==2019.6.16
cffi==1.12.3
cryptography==2.7
Django==2.2.4
djangorestframework==3.10.2
mysqlclient==1.4.4
pycparser==2.19
PyJWT==1.7.1
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
